0829 - home 스크롤 추가 및 구매 페이지 detail 추가
