+++ app/model/cart.py
@@ -3,10 +3,3 @@ from sqlalchemy.ext.declarative import declarative_base
 
 Base = declarative_base()
 
-class Cart(Base):
-    __tablename__ = 'cart'
-    cno = Column(Integer, primary_key=True, autoincrement=True)
-    mno = Column(Integer, ForeignKey('member.mno'), nullable=False)
-    prdno = Column(Integer, ForeignKey('product.prdno'), nullable=False)
-    qty = Column(Integer, nullable=False)
-    price = Column(Integer, nullable=False)
