+++ app/model/product.py
@@ -1,4 +1,6 @@
 from sqlalchemy import Column, Integer, String, Float, Text, DateTime, ForeignKey
+from sqlalchemy.orm import relationship
+
 from app.model.base import Base
 from datetime import datetime
 
@@ -10,6 +12,8 @@ class Product(Base):
     type = Column(String(50), nullable=False)
     qty = Column(Integer, nullable=False, default=0)
     regdate = Column(DateTime, default=datetime.now, nullable=True)
+    # Relationship with PrdAttach
+    attachments = relationship("PrdAttach", back_populates="product")
 
 class PrdAttach(Base):
     __tablename__ = 'prdattach'
@@ -19,3 +23,14 @@ class PrdAttach(Base):
     img2 = Column(String(50), nullable=False)
     img3 = Column(String(50), nullable=False)
     img4 = Column(String(50), nullable=False)
+    # Relationship with Product
+    product = relationship("Product", back_populates="attachments")
+
+
+class Cart(Base):
+    __tablename__ = 'cart'
+    cno = Column(Integer, primary_key=True, autoincrement=True)
+    mno = Column(Integer, ForeignKey('member.mno'), nullable=False)
+    prdno = Column(Integer, ForeignKey('product.prdno'), nullable=False)
+    qty = Column(Integer, nullable=False)
+    price = Column(Integer, nullable=False)
