+++ app/routes/shop.py
@@ -1,6 +1,9 @@
-from fastapi import APIRouter, Request
+from fastapi import APIRouter, Depends, Request, HTTPException
+from sqlalchemy.orm import Session
 from starlette.responses import HTMLResponse
 from starlette.templating import Jinja2Templates
+from app.dbfactory import get_db
+from app.model.product import Product as ProductModel
 
 menu_router = APIRouter()
 templates = Jinja2Templates(directory='views/templates')
@@ -13,11 +16,12 @@ shop_router = APIRouter()
 # ex)
 # member/product/{prdno}
 # admin/product/{prdno}
+@shop_router.get("/item_gallery", response_class=HTMLResponse)
+async def item_gallery(request: Request):
+    return templates.TemplateResponse("shop/item_gallery.html", {"request": request})
+
+
 @shop_router.get("/item_detail", response_class=HTMLResponse)
 async def item_detail(request: Request):
     return templates.TemplateResponse("shop/item_detail.html", {"request": request})
 
-
-@shop_router.get("/item_gallery", response_class=HTMLResponse)
-async def item_gallery(request: Request):
-    return templates.TemplateResponse("shop/item_gallery.html", {"request": request})
