+++ main.py
@@ -9,7 +9,6 @@ from starlette.templating import Jinja2Templates
 
 from app.dbfactory import db_startup, db_shutdown
 from app.routes.admin import admin_router
-from app.routes.cart import cart_router
 from app.routes.member import member_router
 from app.routes.menu import menu_router
 from app.routes.order import order_router
@@ -32,7 +31,6 @@ app.include_router(member_router, prefix='/member')
 app.include_router(menu_router, prefix="/menu")
 app.include_router(shop_router, prefix='/shop')
 app.include_router(order_router, prefix='/order')
-app.include_router(cart_router, prefix='/cart')
 app.include_router(admin_router, prefix='/admin')
 
 @app.get("/", response_class=HTMLResponse)
