+++ views/templates/include/base.html
@@ -7,7 +7,7 @@
     {% block style %}
     <style>
         body {
-        /*{#background-image: url('/static/img/showroom.png');#}*/
+        /*{#background-image: url('/static/img/showroom1.png');#}*/
             background-size: cover; /* 이미지가 화면을 채우도록 설정 */
             background-position: center; /* 이미지가 중앙에 위치하도록 설정 */
             background-repeat: no-repeat; /* 이미지 반복 방지 */
