+++ views/templates/include/header.html
@@ -54,7 +54,7 @@
     <nav>
         <ul class="nav">
             <li class="nav-item"><a class="nav-link active" href="/">Home</a></li>
-            <li class="nav-item"><a class="nav-link active" href="/menu/shop">Shop</a></li>
+            <li class="nav-item"><a class="nav-link active" href="/shop/item_gallery">Shop</a></li>
             <li class="nav-item"><a class="nav-link active" href="/menu/company">Company</a></li>
             <li class="nav-item"><a class="nav-link active" href="/menu/welcome">Welcome</a></li>
         </ul>
