+++ views/templates/index.html
@@ -3,44 +3,44 @@
 <head>
     <meta charset="UTF-8">
     <title>main</title>
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.11.3/font/bootstrap-icons.min.css">
     <link rel="stylesheet" href="/static/css/bs5/bootstrap.min.css">
     <style>
-        /* 배경 이미지 설정 */
+        /* 기존 CSS 코드 그대로 유지 */
         body {
             background-image: url('/static/img/showroom.png');
-            background-size: cover; /* 이미지가 화면을 채우도록 설정 */
-            background-position: center; /* 이미지가 중앙에 위치하도록 설정 */
-            background-repeat: no-repeat; /* 이미지 반복 방지 */
-            min-height: 100vh; /* 화면의 전체 높이를 최소로 설정 */
+            background-size: cover;
+            background-position: center;
+            background-repeat: no-repeat;
+            min-height: 100vh;
             margin: 0;
             padding: 0;
             display: flex;
             flex-direction: column;
             justify-content: space-between;
-            background-color: #f5f5dc; /* 전체 페이지 배경색 */
+            background-color: #f5f5dc;
         }
 
-        /* 헤더와 푸터의 배경을 투명하게 설정 */
         header {
             position: relative;
-            height: 100px; /* 헤더의 높이 설정 */
+            height: 100px;
             display: flex;
             justify-content: space-between;
             align-items: center;
-            padding: 0 20px; /* 헤더 양옆에 패딩 추가 */
-            z-index: 1000; /* 헤더를 배경 이미지 위에 배치 */
+            padding: 0 20px;
+            z-index: 1000;
             color: white;
         }
 
         header img {
-            width: 130px; /* 로고의 너비 설정 */
-            height: auto; /* 로고의 비율 유지 */
+            width: 130px;
+            height: auto;
             margin-top: 20px;
-            margin-right: 20px; /* 로고와 네비게이션 바 사이에 간격 추가 */
+            margin-right: 20px;
         }
 
         nav {
-            flex-grow: 1; /* 네비게이션 바가 로고 옆에 붙고, 공간을 차지하도록 설정 */
+            flex-grow: 1;
         }
 
         .nav {
@@ -63,21 +63,29 @@
         .auth-links {
             display: flex;
             align-items: center;
+            color: black;
         }
 
         .auth-links a {
             text-decoration: none;
             margin-left: 10px;
             color: white;
-            font-size: 15px; /* 글자 크기 설정 */
+            font-size: 15px;
         }
 
-        /* 푸터를 하단에 고정, 반투명 배경색 높이 감소 */
         footer {
             text-align: center;
             padding: 10px 0;
             color: white;
-            margin-top: auto; /* 푸터를 페이지 하단에 위치 */
+            margin-top: auto;
+        }
+
+        h2 {
+            text-align: center;
+            color: white;
+            padding: 300px;
+            font-size: 50px;
+            font-family: "Courier New";
         }
 
         footer p {
@@ -86,10 +94,8 @@
     </style>
 </head>
 <body>
-
 <header class="py-3">
     <h1>KOCAP</h1>
-    {# <img src="/static/img/kocap.png" alt="코 캡 가구 쇼핑몰"> #}
     <nav>
         <ul class="nav">
             <li class="nav-item"><a class="nav-link active" href="/">Home</a></li>
@@ -98,21 +104,42 @@
             <li class="nav-item"><a class="nav-link active" href="/menu/welcome">Welcome</a></li>
         </ul>
     </nav>
-    <!-- 우측 상단의 회원가입 및 로그인 링크 -->
     <div class="auth-links">
         {% if 'logined_uid' not in request.session %}
-        <a href="/member/join">Create an account</a>
-        <a href="/member/login">Sign in</a>
+            <a href="/member/join">Create an account</a>
+            <a href="/member/login">Sign in</a>
         {% else %}
-        <a href="/member/logout">로그아웃</a>
+            <a href="/member/logout">Logout</a>
         {% endif %}
     </div>
 </header>
+
+<!-- 테스트 콘텐츠 추가 -->
+<main>
+    <div style="height: 800px;"> <!-- 충분히 큰 높이로 설정 -->
+        <h2><i class="bi bi-arrow-down"></i>Scroll down<i class="bi bi-arrow-down"></i></h2>
+    </div>
+</main>
 
 <footer>
     <hr>
     <p>copyright &copy; 2024. kocap. all rights reserved.</p>
 </footer>
 
+<script>
+    window.addEventListener('scroll', function() {
+        var scrollPosition = window.scrollY + window.innerHeight;
+        var documentHeight = document.documentElement.scrollHeight;
+
+        // Debug log
+        console.log('Scroll Position:', scrollPosition);
+        console.log('Document Height:', documentHeight);
+
+        if (scrollPosition >= documentHeight * 0.9) {
+            console.log('Redirecting to /shop/item_gallery');
+            window.location.href = '/shop/item_gallery';
+        }
+    });
+</script>
 </body>
 </html>
