+++ views/templates/shop/cart.html
@@ -1,0 +1,95 @@
+{% extends 'include/base.html' %}
+
+{% block title %}
+    장바구니
+{% endblock %}
+
+{% block style %}
+    <style>
+        body {
+            font-family: Arial, sans-serif;
+            margin: 0;
+            padding: 0;
+            background-color: white;
+        }
+        .container {
+            width: 80%;
+            margin: 0 auto;
+            padding: 20px;
+        }
+        .cart-detail {
+            margin-bottom: 20px;
+        }
+        .cart-item {
+            display: flex;
+            flex-wrap: wrap;
+            background-color: #f5f5dc;
+            padding: 20px;
+            margin-bottom: 20px;
+            border-radius: 10px;
+            box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
+        }
+        .cart-item img {
+            flex: 1;
+            max-width: 150px;
+            height: auto;
+            border-radius: 10px;
+            box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
+        }
+        .cart-info {
+            flex: 2;
+            padding-left: 30px;
+        }
+        .cart-title {
+            font-size: 1.8em;
+            margin-bottom: 10px;
+        }
+        .cart-price, .cart-qty {
+            font-size: 1.2em;
+            color: #333;
+            margin-bottom: 10px;
+        }
+        .cart-total {
+            text-align: right;
+            font-size: 1.5em;
+            color: #333;
+            margin-top: 30px;
+        }
+        .checkout-button {
+            display: inline-block;
+            padding: 15px 30px;
+            background-color: #333;
+            color: white;
+            text-decoration: none;
+            font-size: 1.2em;
+            border-radius: 5px;
+            transition: background-color 0.3s;
+        }
+        .checkout-button:hover {
+            background-color: #555;
+        }
+    </style>
+{% endblock %}
+
+{% block main %}
+    <main>
+        <div class="container">
+            <h1>장바구니</h1>
+            {% for item in cart_items %}
+                <div class="cart-item">
+                    <img src="{{ item.product.image_url }}" alt="{{ item.product.prdname }}">
+                    <div class="cart-info">
+                        <h2 class="cart-title">{{ item.product.prdname }}</h2>
+                        <p class="cart-price">가격: ₩{{ item.price }}</p>
+                        <p class="cart-qty">수량: {{ item.qty }}</p>
+                        <p class="cart-total-item">합계: ₩{{ item.price * item.qty }}</p>
+                    </div>
+                </div>
+            {% endfor %}
+            <div class="cart-total">
+                <strong>총 합계: ₩{{ total_price }}</strong>
+            </div>
+            <a href="/checkout" class="checkout-button">결제하기</a>
+        </div>
+    </main>
+{% endblock %}
